backend/requirement.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
openai
python-dotenv
google-generativeai
supabase
pyjwt